Initialize and organize template and pseudocode
Implemented the "first react app" project code as well as the To-Do
List project code as templates for this app.

Moved pseudocode that was originally written in Obsidian to the App
module.

Added some TODO comments as well as other general "brain dump" comments
to the App module related to future ideas/implementations.