Add disabled status to Edit and Submit buttons
Added the "disabled" attribute to both the Edit and Submit buttons to
set up how they will operate in the future. When the app opens, the
Edit button will be disabled. Once all input is complete on the CV form
and the user interacts with the Submit button, the Edit button will be
enabled after the form inputs are shown in HTML format. The Submit
button will then be disabled.

Added a comment reflecting this behavior as a reminder.

Revised the second TODO comment about Rendering Techniques that is
related to the Submit button logic.