Resolve import errors across component files
We kept getting pre-transform errors and internal server errors from
Vite when trying to load the project to our local host on the browser.
The page would just show blank with none of the form additions we made
to the original React/Vite template for our project. Changed all of
the import typos in the JSX component files and the src tag in the
script element of the HTML file so Vite could find the existing files
when loading the SPA in the browser.

SPA now shows up correctly on local host.