Add React prop types rule per TOP recommendations
Implemented a React prop types rule in the ESLint config file and set
it to 'off' based on TOP recommendations from the Rendering Techniques
lesson. This is to prevent the "Missing in props validation" ESLint
warning from showing up as opposed to just ignoring it when it does
pop up.