Reorganize form states with two new components
Implemented conditional rendering into the app. Split up the logic into
two new separate JSX components instead of having everything under the
App component. Updated the button elements in each component to reflect
the change since both buttons were still looking for the handle
functions for their respective functionality on button click. Removed
the onClick method for the button that is to remain disabled depending
on the state of the form. Moved the handle functions responsible for
each enabled button into their respective components.

In the dev comments, I renamed the variable that declared the array
used for reading the form as complete or incomplete on Submit and Edit
functions & buttons. I also renamed the useState example to reflect
this change.