Add preventDefault and sharing state logic
Implemented code into both the Incomplete and Complete component calls
that are inside the App and Incomplete components. This was added based
on reviewing the "Sharing State Between Components" React documentation.
The idea is to change the state of the form between incomplete and
complete using what I read in the doc.

There is a new issue however. For this to work, I will very likely need
to put both handle function that were split up into the new Incomplete
and Complete components back into the App component. Otherwise, I
probably won't be able to "lift the form state" up to the parent App
component, which is required for this project.

Added the preventDefault code to both handle button functions. Updated
the wording on both of the TODO comments at the top of the Incomplete
and Complete components but will probably have to move these comments
back to the App component. Implemented the resumeForm useState in the
App component as well as the formState array to be used alongside it.
Updated the "Choosing the State Structure" dev comment under DOCS.