Rebuild the entire structure of the App
We introduced a brand new file system and approach to how the app is
organized. We rebuilt a different version of the app from the ground up
where we declare multiple states within each section of the app form.
The components are now split up into three sections representing each
section of the CV application form: Personal Info, Work Experience and
Education. Each component has multiple states declared for each form
input within that component or section. New component files were
created and code was implemented from a separate sandbox that we were
working with for this new version.

We did this because we were running into snags trying to brute force
the "one state for the entire form" approach. It was difficult to
figure out how the Submit and Edit button logic would work and how it
would "talk to" the rest of the form when interacted with when it came
to changing the state of the form from "Incomplete" to "Complete".
Displaying the form elements as HTML when submitted was also becoming
an issue, albeit a smaller one (we probably could've figured that out
eventually).