Add dev comment for next steps on Submit button
After getting a clue from Gemini, we went back to the "Sharing State
Between Components" React doc about lifting state up to the App. We
were halfway there with the 'submitted' status, but needed to pass it
down as a prop to our other components and render those components
directly. Next steps will be to implement what we learned based on
Steps 1 and 2 examples in the docs. Added a TODO comment on all the
components as a reminder to try this next.