Modify code in an attempt to debug Submit error
Made the first attempt to debug the errors showing up when the form is
submitted. All info in the label and input sections should show as
heading and <p> tag HTML elements respectively upon Submit.

Used the Personal component as an experiment first by adding the
"submitted" status from the App component as a prop in the function
since that state is already lifted up in the App (parent) component.
Passed the prop down to a condition if/else statement but it didn't
work. The conditional is currently being completely ignored and the
form elements still render when the Personal component is called in App
even after submitting the resume via the Submit button.

Will need to use DevTools and step through each part of the code to
find out why the conditional "if" part isn't rendering and being
skipped over instead.