Modify the code with even more debugging attempts
Attempted to run a separate function that rendered the completed inputs
as HTML via the 'onSubmit' method. Didn't work. Attempted to change
the submitted state in the App component to an empty object, then add
each section as standalone values to setSubmitted on the handleSubmit
function. Didn't work. Attempted to run the same separate function
without the onSubmit method and simply as a conditional, but nothing
renders on the page when opened.

One attempt (#3) does work, but it means that the entered inputs
automatically become HTML as soon as the section is completed, leaving
no need for a Submit button at all. We CAN do this, but its not what
the assignment calls for.