Reimplement conditional if/else attempt logic
The third attempt we made at getting the code to work for the app did
work, but it doesn't align with the assignment instructions. There is
no Edit button needed, which the assignment calls for, because the
HTML automatically generates as sections are completed. However, we will
be implementing the third attempt since our fourth attempt failed and
we are needing a MVP due to time constraints. Added the if/else logic
to the other components (Experience and Education), resolving header
duplication and a small bug related to boolean operators along the way.

Currently there's an issue with text input sections auto generating to
HTML when even a single character is entered into them by users. We'll
be implementing a setTimeout function in each component to deal with
that, giving users time to type in what they need in those text inputs.