Refactor code eliminating all components but App
Since keeping each form section in separate components isn't working,
we decided to move all the code from each component under the App
component, giving the App component direct access to the functions
representing each section as children. We initially wrote this version
of the code in a sandbox first, but it hasn't been tested yet.

Currently, it can't be tested because every useState representing each
form input in the code isn't being read, despite seemingly showing that
the child functions still read each one. Next step is to review the
"Sharing State Between Components" React doc again to find out why this
is the case and if there's a syntax error anywhere. Other options
include stepping through the code in DevTools to see if the console
comes up with anything and reviewing all React docs that have form
inputs as code examples ("State: A Component's Memory", "State as a
Snapshot", etc.).