Transfer all components into the App module
Moved all the components that were in their own respective modules
into the App module due to the app not being able to read the different
states for each input section of the form. Having them all under one
module allows this to happen. We did try several other methods in the
past before deciding on going this path but none of them worked
(exporting with dot notation, setTimeout rendering to delay key
inputs in the text fields, importing useState code directly into other
modules, different onSubmit and onClick call methods as component props,
passing submitted data to other components for displaying them on the
page, etc).

Attempted to get the form data to render as HTML using the ternary
operator method with the "submitted" status, but currently all that
happens is that the form clears and no HTML ever shows. When everything
was in separate modules, we were able to get the HTML to show but there
were bugs (auto populated after one key press, HTML would display
incorrectly, etc). This was all done in one Form component that lifted
all useState methods up instead of keeping the form sections in their
own separate components due to some useStates not being recognized.