Revert back to separate components per section
After reading "Thinking in React" and reviewing other approaches to
React forms, I've decided to go back to my previous setup of having a
separate component for each section of the resume form. The idea going
forward will be to have the next section of each component be a
subcomponent of the previous section of the form all the way down to
the Submit and Edit buttons.

At some point, I want to figure out how to do one handle function that
is responsible for all the form input changes, not just a specific
handle function for each input. This may be the next step after I
finish reorganizing (removing code) the Form component inside the App
module.