Revise "Complete" component and add dev comments
Repurposed the "Complete" component we had in the first version of the
app that tried to have one state for the entire form (Incomplete and
Complete states). Instead of a conditional showing the completed HTML
of the form, the component simply shows the now revised HTML that we
set up in a previous iteration and tried to implement in a separate
function.

Added TODO dev comments to each component having to do with what code
is needed if we have to put the buttons in their own subcomponent.