Add a new Buttons component with modifications
Gave the Submit and Edit buttons in the app their own Buttons component
and rendered the buttons there instead. Called the Buttons component
under Education component to fall in line with the new hierarchy we
have for our React components. Called the Complete component under the
Buttons component and passed all states from the other sections
(Personal, Experience, etc.) as props in the Complete function.

Modified the variables embedded in the HTML rendering in the Complete
section as we were getting error messages stating that the values of
the variables weren't being. Since we are now passing all the states as
"props" in the Complete component, we decided to go with dot notation
{props.variable} and the errors have disappeared for now.

There still isn't any HTML rendering on the screen after clicking
Submit, though. We haven't gotten any further. Next steps: read step 5
of "Thinking in React" to see if I missed anything, step through the
code in DevTools one more time and pay very close attention to what's
happening when <Complete /> in the handleSubmit function is called. We
may possibly have to implement empty arrays with map methods to get
things to render. The problem is how to get the buttton to "read" whats
in the array without a CRUD method or how to get the info to display on
one Submit button if the array is empty in the first place (is the info
added to the array upon input/keyup?)