Add props to remaining components in hierarchy
After completing the passing of all props to each component down the
current hierarchy to make sure the Complete component has all the props
from the other form sections, nothing renders at all on initial page
load. I read through step 5 of "Thinking in React" and couldn't come
up with anything. There's currently no code to step through in DevTools,
even when I go into the Components tab

Next steps: revert back to previous version of the code before going
down the JayBee rabbit hole (he did ultimately say the approach was
wrong) if possible, step through the code in DevTools and pay very
close attention to what's happening when the handleSubmit function is
called. We may possibly have to implement empty arrays with map methods
to get things to render. The problem is how to get the buttton to
"read" whats in the array without a CRUD method or how to get the info
to display on one Submit button if the array is empty in the first
place (is the info added to the array upon input/keyup?).