Refactor all input states to the Form component
Moved all states related to form inputs under the Form function in the
App component instead. Called all the subsections of the form (Personal,
Experience, Education, Buttons, Complete) within the Form component.
Brought all the states down as individual props in each respective
component instead of a single "props" argument trying to use dot
notation to represent all of them as this prevented the page from even
rendering on initial load. App now renders again on first page load.

Added a return statement to the handleSubmit function in buttons
thinking this would finally resolve the rendering issue (HTML not
rendering on submit). Modified some of the TODO dev comments in the App
component with docs to go over next time before making another attempt
to render the HTML from the completed inputs.

After going over the docs again, possible next steps are to step through
the code in DevTools and pay very close attention to what's happening
when the handleSubmit function is called. We may not have to implement
empty arrays with map methods to get things to render after all. If we
do go this route, the problem that still remains is how to get the
button to "read" whats in the array without a CRUD method or how to get
the info to display on one Submit button if the array is empty in the
first place (is the info added to the array upon input/keyup?).