Remove modules, clean up code and dev comments
Removed all code and dev comments in all modules that were no longer
needed. Removed the entire Vite + React template except for the logos.
Moved the logos to the top of the application above the heading as a
finishing touch. Removed the Buttons and Incomplete modules as they are
no longer being used. Tested the app to make sure nothing broke after
removing the template. The form, Submit and Edit buttons still work as
intended.

Next steps:
- Deploy to Vercel